refactor: Replaces IDataSource with ICamera.
We want to reduce the scope of simulation within the application, this resulted in only wanted to mock the actual camera interface usage as opposed to the entire data source since we always want to include OpenCV.
To do this, we use an ICamera interface that can make real OpenCV VideoCapture calls, or the MockCamera which is a dummy. Having the mock available allows us to run on hardware without needing to invoke the webcam and also allows us to do easier unit testing locally and in ci/cd envrionments.